Disable NRVO for entry functions and patch constant functions
